# code/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App API",
        "version": "0.1"
    },
    "paths": {
        "/blog": {
            "get": {
                "tags": [
                    "blog"
                ],
                "operationId": "39c2879e2df0589d71832a6e736a64bb",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Paginator"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "blog"
                ],
                "operationId": "631f83b9d4166771e126bac62dda9757",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "author",
                                    "text"
                                ],
                                "properties": {
                                    "author": {
                                        "type": "string"
                                    },
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "new post id",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/page/{page}": {
            "get": {
                "tags": [
                    "blog"
                ],
                "operationId": "6e5be8fe650578e56e35eb9b9c345a3d",
                "parameters": [
                    {
                        "parameter": "page",
                        "name": "page",
                        "in": "path",
                        "description": "blog page",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Paginator"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/{id}": {
            "get": {
                "tags": [
                    "blog"
                ],
                "operationId": "d7fef528af9e76804241c0b7a12567c0",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "blog post id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "blog"
                ],
                "operationId": "9002dc2e531000244dc68062f1862daf",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "blog post id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "author": {
                                        "type": "string"
                                    },
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "auth"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "blog"
                ],
                "operationId": "c502606b8985f69e7a0bf0cfceec2f20",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "blog post id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/blog/{id}/comments": {
            "get": {
                "tags": [
                    "blog"
                ],
                "operationId": "6d38baa83940397beed4de00229ca230",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "blog post id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/comment/{post_id}": {
            "post": {
                "tags": [
                    "comment"
                ],
                "operationId": "47e58038a86ae7b7436eddeaec184b41",
                "parameters": [
                    {
                        "parameter": "post_id",
                        "name": "post_id",
                        "in": "path",
                        "description": "blog post id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "author",
                                    "text"
                                ],
                                "properties": {
                                    "author": {
                                        "type": "string"
                                    },
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "new comment id",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/comment/{id}": {
            "put": {
                "tags": [
                    "comment"
                ],
                "operationId": "18252560ec1cdfe10e0c7d0c1e8dfc3e",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "comment id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "auth"
                    }
                ]
            }
        },
        "/": {
            "get": {
                "operationId": "789aebb8c643aec7a5e6e6b06d2f62ea",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Comment": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "post_id": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "author": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Paginator": {
                "properties": {
                    "total": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer"
                    },
                    "currentPage": {
                        "type": "integer"
                    },
                    "pages": {
                        "type": "integer"
                    },
                    "options": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "auth": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/login",
                        "scopes": {}
                    }
                }
            }
        }
    }
}